+++ packages/portal/spec-renderer/vite.config.ts
@@ -38,10 +38,11 @@ const customPlugins = [
 config.plugins = customPlugins
 
 // If we are trying to preview a build of the local `package/spec-renderer/sandbox` directory,
-// unset the external and lib properties
-if (process.env.PREVIEW_SANDBOX) {
-  config.build.rollupOptions.external = undefined
+// unset the lib, rollupOptions.external and rollupOptions.output.globals properties
+if (process.env.USE_SANDBOX) {
   config.build.lib = undefined
+  config.build.rollupOptions.external = undefined
+  config.build.rollupOptions.output.global = undefined
 }
 
 export default config