+++ packages/core/copy-uuid/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/copy-uuid@1.3.1...@kong-ui-public/copy-uuid@1.3.2) (2023-10-23)
+
+**Note:** Version bump only for package @kong-ui-public/copy-uuid
+
+
+
+
+
 ## [1.3.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/copy-uuid@1.3.0...@kong-ui-public/copy-uuid@1.3.1) (2023-10-18)
 
 **Note:** Version bump only for package @kong-ui-public/copy-uuid