+++ packages/core/error-boundary/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/error-boundary@0.2.1...@kong-ui-public/error-boundary@0.2.2) (2023-10-23)
+
+**Note:** Version bump only for package @kong-ui-public/error-boundary
+
+
+
+
+
+## [0.2.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/error-boundary@0.2.0...@kong-ui-public/error-boundary@0.2.1) (2023-10-19)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @kong/icons to ^1.7.6 ([#846](https://github.com/Kong/public-ui-components/issues/846)) ([b658562](https://github.com/Kong/public-ui-components/commit/b6585621597b9e27ed17aab8f93b58832af48890))
+
+
+
+
+
+# [0.2.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/error-boundary@0.1.1...@kong-ui-public/error-boundary@0.2.0) (2023-10-18)
+
+
+### Features
+
+* **error-boundary:** include data-testid [KHCP-9478] ([#845](https://github.com/Kong/public-ui-components/issues/845)) ([2558acb](https://github.com/Kong/public-ui-components/commit/2558acba506a81756614a48b31e54f1714b8d526))
+
+
+
+
+
 ## [0.1.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/error-boundary@0.1.0...@kong-ui-public/error-boundary@0.1.1) (2023-10-18)
 
 **Note:** Version bump only for package @kong-ui-public/error-boundary