+++ packages/entities/entities-upstreams-targets/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.6](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-upstreams-targets@1.3.5...@kong-ui-public/entities-upstreams-targets@1.3.6) (2023-10-23)
+
+**Note:** Version bump only for package @kong-ui-public/entities-upstreams-targets
+
+
+
+
+
+## [1.3.5](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-upstreams-targets@1.3.4...@kong-ui-public/entities-upstreams-targets@1.3.5) (2023-10-19)
+
+**Note:** Version bump only for package @kong-ui-public/entities-upstreams-targets
+
+
+
+
+
 ## [1.3.4](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-upstreams-targets@1.3.3...@kong-ui-public/entities-upstreams-targets@1.3.4) (2023-10-18)
 
 **Note:** Version bump only for package @kong-ui-public/entities-upstreams-targets