+++ packages/portal/spec-renderer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.11.55](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/spec-renderer@0.11.54...@kong-ui-public/spec-renderer@0.11.55) (2023-10-23)
+
+**Note:** Version bump only for package @kong-ui-public/spec-renderer
+
+
+
+
+
 ## [0.11.54](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/spec-renderer@0.11.53...@kong-ui-public/spec-renderer@0.11.54) (2023-10-18)
 
 **Note:** Version bump only for package @kong-ui-public/spec-renderer